+++ ToYouProject/.gitignore
@@ -1,0 +1,10 @@
+*.iml
+.gradle
+/local.properties
+/.idea/workspace.xml
+/.idea/libraries
+/app/src/main/res/values/ignore_values
+.DS_Store
+/build
+/captures
+.externalNativeBuild
